Move getGroupIdC into BDCS/Groups.hs.
